3 overflow fixes

+++ l2fwd/server.c
@@ -65,7 +65,7 @@ void process_batch_gpu(struct rte_mbuf **pkts, int nb_pkts,
 	struct ipv4_hdr *ip_hdr;
 	uint32_t dst_ip;
 
-	int head = lc_wmq->head;
+	LL head = lc_wmq->head;
 	LL *srv_tsc;
 
 	foreach(batch_index, nb_pkts) {
